pyproject.toml
[tool.poetry]
name = "ragulate"
version = "0.0.1"
description = "A tool for evaluating RAG pipelines"
authors = ["Eric Pinzur <epinzur@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "ragulate/*"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
epinzur-trulens-eval = ">=0.30.1b0"
kaleido = "0.2.1"
inflection = "^0.5.1"
llama-index-core = "^0.10.31"
numpy = ">=1.23.5"
pandas = ">=2.2.2"
plotly = "^5.22.0"
python-dotenv = ">=1.0.0"
tqdm = ">=4.66.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
langchain-astradb = "0.3.3"
langchain-community = "0.0.38"
langchain-core = "0.1.52"
langchain-openai = "0.1.3"

[build-system]
requires = ["poetry-core", "setuptools>=42", "wheel", "pip"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ragulate = "ragulate.cli:main"
